cleared up legacy logging code
