+++ datavis/decode_binary.py
@@ -4,17 +4,8 @@ import sys
 import os
 import math
 
-# Expected columns matching the BinaryLogRecord layout exactly
-# Original 20 fields + 3 raw mocap + 6 IMU = 29 float fields + 1 u32 timestamp
 CSV_HEADER = "ts,x,y,yaw,x_des,y_des,yaw_des,v_ff,w_ff,v_actual,w_actual,omega_l_cmd,omega_r_cmd,omega_l_meas,omega_r_meas,duty_l,duty_r,x_err,y_err,yaw_err,x_raw,y_raw,yaw_raw,acc_x,acc_y,acc_z,gyro_x,gyro_y,gyro_z\n"
-RECORD_FORMAT = "<I28f"
-RECORD_SIZE = struct.calcsize(RECORD_FORMAT)  # 116 bytes
 MAGIC_HEADER = b"\xaa\xbb\xcc\xdd"
-
-# Also support the legacy 80-byte format (20 fields: 1 u32 + 19 f32)
-LEGACY_RECORD_FORMAT = "<I19f"
-LEGACY_RECORD_SIZE = struct.calcsize(LEGACY_RECORD_FORMAT)  # 80 bytes
-LEGACY_CSV_HEADER = "ts,x,y,yaw,x_des,y_des,yaw_des,v_ff,w_ff,v_actual,w_actual,omega_l_cmd,omega_r_cmd,omega_l_meas,omega_r_meas,duty_l,duty_r,x_err,y_err,yaw_err\n"
 
 
 def format_value(val):
@@ -145,41 +136,8 @@ def decode_file(input_path):
             if current_ts is not None:
                 records.append(",".join(current_row))
         else:
-            # Determine record size: try new format first, fall back to legacy
-            if data_size > 0 and data_size % RECORD_SIZE == 0:
-                record_format = RECORD_FORMAT
-                record_size = RECORD_SIZE
-                csv_header = CSV_HEADER
-                print(f"Detected fixed 116-byte format ({record_size} bytes/record, {data_size // record_size} records)")
-            elif data_size > 0 and data_size % LEGACY_RECORD_SIZE == 0:
-                record_format = LEGACY_RECORD_FORMAT
-                record_size = LEGACY_RECORD_SIZE
-                csv_header = LEGACY_CSV_HEADER
-                print(f"Detected legacy 80-byte format ({record_size} bytes/record, {data_size // record_size} records)")
-            else:
-                # Try new format first (may have trailing partial record)
-                record_format = RECORD_FORMAT
-                record_size = RECORD_SIZE
-                csv_header = CSV_HEADER
-                print(f"Warning: Data size {data_size} is not a clean multiple of record size. Trying fixed 116-byte format.")
-                
-            while True:
-                chunk = bin_file.read(record_size)
-                if not chunk:
-                    break
-                if len(chunk) < record_size:
-                    print(f"Warning: Trailing partial record ignored ({len(chunk)} bytes)")
-                    break
-                    
-                # Unpack record
-                unpacked = struct.unpack(record_format, chunk)
-                # unpacked[0] is t_ms (u32), rest are f32
-                # Format: ts is integer, floats preserve NaN as empty
-                row_parts = [str(unpacked[0])]
-                for val in unpacked[1:]:
-                    row_parts.append(format_value(val))
-                row_str = ",".join(row_parts)
-                records.append(row_str)
+            print(f"Error: Unsupported legacy binary format. Only Compact Tagged Binary logs can be decoded.")
+            return False
             
     # Overwrite the original binary file with the CSV content
     print(f"Overwriting binary file with CSV data: {input_path}")
