+++ firmware/src/lib.rs
@@ -43,7 +43,7 @@ macro_rules! debug_warn {
 pub mod button;
 pub mod buzzer;
 pub mod control_types;
-pub mod diffdrive;
+// pub mod diffdrive; //legacy
 pub mod ekf;
 pub mod encoder;
 pub mod encoder_lib;
